requirements.txt
flake8==4.0.1
isort==5.10.1
black==21.10b0

PyGithub==1.55
pyhumps==3.0.2
sqlalchemy==1.4.27
alembic==1.7.5
rich==10.14.0
GitPython==3.1.24
